fix: delete unneeded abstraction, better option name, default serializer
